pathToCppFile: fix for nested source().
This fixes errors that occur when GBS-Chip-Gmatrix.R is sourced from a
script in a different directory which is in turn sourced by another
script.

If multiple source calls are nested, then we should only look at the
path parameter for the inner-most source() frame. If the inner-most
frame had that chdir option set, then we don't need to translate the
filename.