Nix: bump to nixpkgs-unstable and expose required R packages
The previous constraint from gbs-prism to use an older version of
nixpkgs has been removed, and the advantage of using nixpkgs-unstable
is that it has the fixed Rcpp that we need.

Exposing the R packages directly seems now to be required in order to
use KGD both in gbs-prism and downstream of that in gbs-uber-keyfiles.